requirements.txt
aiohttp==3.9.5
aiomysql==0.2.0
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
dateparser==1.2.0
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
loguru==0.7.2
multidict==6.0.5
nextcord==2.6.0
packaging==24.1
pluggy==1.5.0
prometheus_client==0.20.0
pycparser==2.22
PyMySQL==1.1.1
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
setuptools==70.3.0
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
vcrpy==6.0.1
wrapt==1.16.0
yarl==1.9.4
